tools/requirements.txt
wheel
paho-mqtt
matplotlib